return response entity as data for most of the permission rest controller methods

+++ src/main/kotlin/nl/hva/backend/rest/PermissionRestController.kt
@@ -1,18 +1,19 @@
 package nl.hva.backend.rest
 
+import nl.hva.backend.application.api.AccountService
 import nl.hva.backend.application.api.MedicalRecordService
 import nl.hva.backend.application.api.PermissionService
-import nl.hva.backend.application.dto.DiagnosisDTO
-import nl.hva.backend.application.dto.ExerciseDTO
-import nl.hva.backend.application.dto.MedicationDTO
-import nl.hva.backend.application.dto.NoteDTO
+import nl.hva.backend.application.dto.*
 import nl.hva.backend.application.dto.many_to_many.DiagnosisCareProviderDTO
 import nl.hva.backend.application.dto.many_to_many.ExerciseCareProviderDTO
 import nl.hva.backend.application.dto.many_to_many.MedicationCareProviderDTO
 import nl.hva.backend.application.dto.many_to_many.NoteCareProviderDTO
 import nl.hva.backend.domain.ids.*
+import nl.hva.backend.rest.exceptions.NotExistingException
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.format.annotation.DateTimeFormat
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 import java.time.LocalDate
 
@@ -23,6 +24,8 @@ class PermissionRestController {
     @Autowired
     private lateinit var permissionService: PermissionService
 
+    @Autowired
+    private lateinit var accountService: AccountService
 
     @Autowired
     private lateinit var medicalRecordService: MedicalRecordService
@@ -51,22 +54,35 @@ class PermissionRestController {
     fun getMedicationCareProviderRelationById(
         mrId: String,
         cpId: String
-    ): List<MedicationDTO> {
-        //Check all permissions the care provider has
-        val medicationCareProviderDTOs: List<MedicationCareProviderDTO> =
-            this.permissionService.getMedicationCareProviderRelationById(CareProviderId(cpId))
+    ): ResponseEntity<List<MedicationDTO>> {
+        val medicalRecordDTO: List<MedicalRecordDTO> = this.medicalRecordService.getMedicalRecord(MedicalRecordId(mrId))
 
-        //Return only the medication of the requested patient
-        val medicationDTOs: ArrayList<MedicationDTO> = arrayListOf()
-        for (medicationCareProviderDTO in medicationCareProviderDTOs) {
-            val medicationDTO: List<MedicationDTO> = this.medicalRecordService.getMedicationByIdAndMedicalRecordId(
-                MedicationId(medicationCareProviderDTO.medId()),
-                MedicalRecordId(mrId)
-            )
-            medicationDTOs.add(medicationDTO[0])
+        if (medicalRecordDTO.isNotEmpty()) {
+            val careProviderDTO: List<CareProviderDTO> = this.accountService.getCareProviderById(CareProviderId(cpId))
+
+            if (careProviderDTO.isNotEmpty()) {
+                //Check all permissions the care provider has
+                val medicationCareProviderDTOs: List<MedicationCareProviderDTO> =
+                    this.permissionService.getMedicationCareProviderRelationById(CareProviderId(cpId))
+
+                //Return only the medication of the requested patient
+                val medicationDTOs: ArrayList<MedicationDTO> = arrayListOf()
+                for (medicationCareProviderDTO in medicationCareProviderDTOs) {
+                    val medicationDTO: List<MedicationDTO> =
+                        this.medicalRecordService.getMedicationByIdAndMedicalRecordId(
+                            MedicationId(medicationCareProviderDTO.medId()),
+                            MedicalRecordId(mrId)
+                        )
+                    medicationDTOs.add(medicationDTO[0])
+                }
+
+                return ResponseEntity.status(HttpStatus.OK).body(medicationDTOs)
+            } else {
+                throw NotExistingException("Care provider with id \'$cpId\' does not exist.")
+            }
+        } else {
+            throw NotExistingException("Medical record with id \'$mrId\' does not exist.")
         }
-
-        return medicationDTOs
     }
 
     @PostMapping("/createMedicationPermission")
@@ -75,23 +91,51 @@ class PermissionRestController {
         careProviderId: String,
         @RequestParam("validDate")
         @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) validDate: LocalDate
-    ) {
-        this.permissionService.createPermissionLinkMedication(
-            MedicationId(medicationId),
-            CareProviderId(careProviderId),
-            validDate
-        )
+    ): ResponseEntity<String> {
+        val medicationDTO: List<MedicationDTO> = this.medicalRecordService.getMedicationById(MedicationId(medicationId))
+
+        if (medicationDTO.isNotEmpty()) {
+            val careProviderDTO: List<CareProviderDTO> = this.accountService.getCareProviderById(CareProviderId(careProviderId))
+
+            if (careProviderDTO.isNotEmpty()) {
+                this.permissionService.createPermissionLinkMedication(
+                    MedicationId(medicationId),
+                    CareProviderId(careProviderId),
+                    validDate
+                )
+
+                return ResponseEntity.status(HttpStatus.OK).body("New permission for care provider with id \'$careProviderId\' for medication with id \'$medicationId\' successfully created.")
+            } else {
+                throw NotExistingException("Care provider with id \'$careProviderId\' does not exist.")
+            }
+        } else {
+            throw NotExistingException("Medication with id \'$medicationId\' does not exist.")
+        }
     }
 
     @DeleteMapping("/removeMedicationPermission")
     fun deleteMedicationPermission(
         medicationId: String,
         careProviderId: String
-    ) {
-        this.permissionService.removeSelectedMedicationPermission(
-            MedicationId(medicationId),
-            CareProviderId(careProviderId)
-        )
+    ): ResponseEntity<String> {
+        val medicationDTO: List<MedicationDTO> = this.medicalRecordService.getMedicationById(MedicationId(medicationId))
+
+        if (medicationDTO.isNotEmpty()) {
+            val careProviderDTO: List<CareProviderDTO> = this.accountService.getCareProviderById(CareProviderId(careProviderId))
+
+            if (careProviderDTO.isNotEmpty()) {
+                this.permissionService.removeSelectedMedicationPermission(
+                    MedicationId(medicationId),
+                    CareProviderId(careProviderId)
+                )
+
+                return ResponseEntity.status(HttpStatus.OK).body("Permission for care provider with id \'$careProviderId\' for medication with id \'$medicationId\' successfully removed.")
+            } else {
+                throw NotExistingException("Care provider with id \'$careProviderId\' does not exist.")
+            }
+        } else {
+            throw NotExistingException("Medication with id \'$medicationId\' does not exist.")
+        }
     }
 
 
@@ -104,24 +148,36 @@ class PermissionRestController {
     fun getNoteCareProviderRelationById(
         mrId: String,
         cpId: String
-    ): List<NoteDTO> {
-        //Check all permissions the care provider has
-        val noteCareProviderDTOs: List<NoteCareProviderDTO> =
-            this.permissionService.getNoteCareProviderRelationById(
-                CareProviderId(cpId)
-            )
+    ): ResponseEntity<List<NoteDTO>> {
+        val medicalRecordDTO: List<MedicalRecordDTO> = this.medicalRecordService.getMedicalRecord(MedicalRecordId(mrId))
 
-        //Return only the note of the requested patient
-        val noteDTOs: ArrayList<NoteDTO> = arrayListOf()
-        for (noteCareProviderDTO in noteCareProviderDTOs) {
-            val noteDTO: List<NoteDTO> = this.medicalRecordService.getNoteByIdAndMedicalRecordId(
-                NoteId(noteCareProviderDTO.noteId()),
-                MedicalRecordId(mrId)
-            )
-            noteDTOs.add(noteDTO[0])
+        if (medicalRecordDTO.isNotEmpty()) {
+            val careProviderDTO: List<CareProviderDTO> = this.accountService.getCareProviderById(CareProviderId(cpId))
+
+            if (careProviderDTO.isNotEmpty()) {
+                //Check all permissions the care provider has
+                val noteCareProviderDTOs: List<NoteCareProviderDTO> =
+                    this.permissionService.getNoteCareProviderRelationById(
+                        CareProviderId(cpId)
+                    )
+
+                //Return only the note of the requested patient
+                val noteDTOs: ArrayList<NoteDTO> = arrayListOf()
+                for (noteCareProviderDTO in noteCareProviderDTOs) {
+                    val noteDTO: List<NoteDTO> = this.medicalRecordService.getNoteByIdAndMedicalRecordId(
+                        NoteId(noteCareProviderDTO.noteId()),
+                        MedicalRecordId(mrId)
+                    )
+                    noteDTOs.add(noteDTO[0])
+                }
+
+                return ResponseEntity.status(HttpStatus.OK).body(noteDTOs)
+            } else {
+                throw NotExistingException("Care provider with id \'$cpId\' does not exist.")
+            }
+        } else {
+            throw NotExistingException("Medical record with id \'$mrId\' does not exist.")
         }
-
-        return noteDTOs
     }
 
     @PostMapping("/createNotePermission")
@@ -159,24 +215,36 @@ class PermissionRestController {
     fun getDiagnosisCareProviderRelationById(
         mrId: String,
         cpId: String
-    ): List<DiagnosisDTO> {
-        //Check all permissions the care provider has
-        val diagnosisCareProviderDTOs: List<DiagnosisCareProviderDTO> =
-            this.permissionService.getDiagnosisCareProviderRelationById(
-                CareProviderId(cpId)
-            )
+    ): ResponseEntity<List<DiagnosisDTO>> {
+        val medicalRecordDTO: List<MedicalRecordDTO> = this.medicalRecordService.getMedicalRecord(MedicalRecordId(mrId))
 
-        //Return only the medication of the requested patient
-        val diagnosisDTOs: ArrayList<DiagnosisDTO> = arrayListOf()
-        for (noteCareProviderDTO in diagnosisCareProviderDTOs) {
-            val diagnosisDTO: List<DiagnosisDTO> = this.medicalRecordService.getDiagnosisByIdAndMedicalRecordId(
-                DiagnosisId(noteCareProviderDTO.diagId()),
-                MedicalRecordId(mrId)
-            )
-            diagnosisDTOs.add(diagnosisDTO[0])
+        if (medicalRecordDTO.isNotEmpty()) {
+            val careProviderDTO: List<CareProviderDTO> = this.accountService.getCareProviderById(CareProviderId(cpId))
+
+            if (careProviderDTO.isNotEmpty()) {
+                //Check all permissions the care provider has
+                val diagnosisCareProviderDTOs: List<DiagnosisCareProviderDTO> =
+                    this.permissionService.getDiagnosisCareProviderRelationById(
+                        CareProviderId(cpId)
+                    )
+
+                //Return only the medication of the requested patient
+                val diagnosisDTOs: ArrayList<DiagnosisDTO> = arrayListOf()
+                for (noteCareProviderDTO in diagnosisCareProviderDTOs) {
+                    val diagnosisDTO: List<DiagnosisDTO> = this.medicalRecordService.getDiagnosisByIdAndMedicalRecordId(
+                        DiagnosisId(noteCareProviderDTO.diagId()),
+                        MedicalRecordId(mrId)
+                    )
+                    diagnosisDTOs.add(diagnosisDTO[0])
+                }
+
+                return ResponseEntity.status(HttpStatus.OK).body(diagnosisDTOs)
+            } else {
+                throw NotExistingException("Care provider with id \'$cpId\' does not exist.")
+            }
+        } else {
+            throw NotExistingException("Medical record with id \'$mrId\' does not exist.")
         }
-
-        return diagnosisDTOs
     }
 
     @PostMapping("/createDiagnosisPermission")
@@ -214,24 +282,37 @@ class PermissionRestController {
     fun getExerciseCareProviderRelationById(
         mrId: String,
         cpId: String
-    ): List<ExerciseDTO> {
-        //Check all permissions the care provider has
-        val exerciseCareProviderDTOs: List<ExerciseCareProviderDTO> =
-            this.permissionService.getExerciseCareProviderRelationById(
-                CareProviderId(cpId)
-            )
+    ): ResponseEntity<List<ExerciseDTO>> {
+        val medicalRecordDTO: List<MedicalRecordDTO> = this.medicalRecordService.getMedicalRecord(MedicalRecordId(mrId))
 
-        //Return only the medication of the requested patient
-        val exerciseDTOs: ArrayList<ExerciseDTO> = arrayListOf()
-        for (exerciseCareProviderDTO in exerciseCareProviderDTOs) {
-            val exerciseDTO: List<ExerciseDTO> = this.medicalRecordService.getExerciseByIdAndMedicalRecordId(
-                ExerciseId(exerciseCareProviderDTO.exerId()),
-                MedicalRecordId(mrId)
-            )
-            exerciseDTOs.add(exerciseDTO[0])
+        if (medicalRecordDTO.isNotEmpty()) {
+            val careProviderDTO: List<CareProviderDTO> = this.accountService.getCareProviderById(CareProviderId(cpId))
+
+            if (careProviderDTO.isNotEmpty()) {
+                //Check all permissions the care provider has
+                val exerciseCareProviderDTOs: List<ExerciseCareProviderDTO> =
+                    this.permissionService.getExerciseCareProviderRelationById(
+                        CareProviderId(cpId)
+                    )
+
+                //Return only the medication of the requested patient
+                val exerciseDTOs: ArrayList<ExerciseDTO> = arrayListOf()
+                for (exerciseCareProviderDTO in exerciseCareProviderDTOs) {
+                    val exerciseDTO: List<ExerciseDTO> = this.medicalRecordService.getExerciseByIdAndMedicalRecordId(
+                        ExerciseId(exerciseCareProviderDTO.exerId()),
+                        MedicalRecordId(mrId)
+                    )
+                    exerciseDTOs.add(exerciseDTO[0])
+                }
+
+                return ResponseEntity.status(HttpStatus.OK).body(exerciseDTOs)
+
+            } else {
+                throw NotExistingException("Care provider with id \'$cpId\' does not exist.")
+            }
+        } else {
+            throw NotExistingException("Medical record with id \'$mrId\' does not exist.")
         }
-
-        return exerciseDTOs
     }
 
     @PostMapping("/createExercisePermission")
